add impersonation logic
This provides a unified mechanism for determining the effective identity
of a calling application (the logic was ported from gm-fabric-go).

The idea is that a service can use USER_DN and EXTERNAL_SYS_DN headers
as supplied by a proxy, and as a fallback it can pull the DN from the
peers cert (assuming a TLS connection with a provided client cert).